Phase-5/backend/requirements.txt
fastapi
sqlmodel
sqlalchemy
asyncpg
uvicorn[standard]
pydantic
pydantic-settings
alembic
python-jose[cryptography]
passlib[argon2]
python-multipart
httpx
psycopg2-binary
pytest
pytest-asyncio
pytest-cov
black
isort
flake8
python-dotenv
bcrypt
pyjwt
mcp
openai-agents
dapr
dapr-ext-fastapi
python-dateutil